Changed Tear and talisman textures
I always thought the texture didn't correspond to a "tear" so I made another (imho better looking) texture that is a derivative of a ghast tear (It also glows, yay). The amulet colors were remade to match that change. A few rendering changes had to be done to render the tear nicely in hand. Thanks to TechnicianLP for that.